Add video device & monitor; assembler, UI, and microcode revamp
Introduce a memory-mapped video device and ImGui-based video monitor supporting ASCII and RGBA modes.
Refactor assembler and instruction encoding for width/size keywords, string literals, and improved operand handling.
Enhance UI with new MemoryView, jump dialog, register/ALU display, and menu bar.
Add NOP microcode, improve stack/memory instructions, and update instruction flags.
Include sample programs for video output. Update dependencies and enable native debugging.

+++ CPUSimulator/Core/Assembly/Lexical/Token.cs
@@ -117,7 +117,7 @@
 
         public readonly StringSpan AsStringSpan()
         {
-            return new(Text, (int)Length);
+            return new(Text, Length);
         }
 
         public override readonly string ToString()
